renames to favourites_view. first working version
